+++ src/types.ts
@@ -3,13 +3,6 @@
  */
 export type Await<T> = T | Promise<T> | PromiseLike<T>;
 
-/**
- * Basic error handler.
- * @param error The error thrown
- * @param jwt The decoded token. May or may not be available.
- */
-export type ErrorHandler<T, K extends keyof Properties<T> = never> = (error: Error | any, jwt?: JWT<T, K>) => any;
-
 type MethodNames<T> = { [P in keyof T]: T[P] extends (...args: any[]) => any ? P : never; }[keyof T];
 
 /**
@@ -38,7 +31,7 @@ export type FindSubjectFunction<A extends any[], T = never, K extends keyof Prop
 export type VerifySubjectFunction = (subject: string) => Await<boolean | undefined>;
 
 /**
- * Generate an unique identifier for token.
+ * Generate an unique identifier for JWT token.
  * @returns unique identifier.
  */
 export type GenerateIDFunction = () => Await<string>;
@@ -113,10 +106,6 @@ export interface JWTManagerOptions<A extends any[], T = never, K extends keyof P
    */
   clockTolerance?: number;
   /**
-   * Handle errors.
-   */
-  onError?: ErrorHandler<T, K>;
-  /**
    * Either the secret for HMAC algorithms, or the PEM encoded public key for RSA and ECDSA.
    */
   secretOrPublicKey?: string | Buffer;
@@ -137,7 +126,7 @@ export interface JWTManagerOptions<A extends any[], T = never, K extends keyof P
    */
   verifySubject?: VerifySubjectFunction;
   /**
-   * Generate an unique identifier for token.
+   * Generate an unique identifier for JWT token.
    */
   generateID: GenerateIDFunction;
 }